Update overflow settings and implement goToTop function
The body's overflow property in the styles.scss file was modified to overflow-x. In footer.component.ts, the dependency was injected to access the DOCUMENT object. Using this, the goToTop function was modified to scroll the page to the top. Additionally, some redundant CSS was removed from app.component.scss.